+++ go-api/models/models.go
@@ -1,8 +1,8 @@
 package models
 
 type Book struct {
-	BookId    int64  `json:"bookid"`
-	Name      string `json:"name"`
-	Price     int64  `json:"price"`
-	Publisher string `json:"publisher"`
+	BookId    int64   `json:"bookid"`
+	Name      string  `json:"name"`
+	Price     float64 `json:"price"`
+	Publisher string  `json:"publisher"`
 }
